Moving criteriaList and hangingTypes to TeamLab
