+++ app/src/main/java/com/bignerdranch/android/criminalintent/HangFragment.java
@@ -42,7 +42,7 @@ public class HangFragment extends DialogFragment {
         mHangBar.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
             @Override
             public void onProgressChanged(SeekBar seekBar, int progress, boolean b) {
-                Toast.makeText(getActivity(), Team.hangingTypes.get(progress), Toast.LENGTH_SHORT).show();
+                Toast.makeText(getActivity(), TeamLab.hangingTypes.get(progress), Toast.LENGTH_SHORT).show();
             }
 
             @Override
